Added a whole bunch of code to import an optionsxpress file and read it into a stock account object.  It's *mostly* working with some exceptions.  There's some kind of bug around getting the $set->startDate() for some sets (the first one being ESI in the test file).  I have a good data set now, as I downloaded my entire activity file and tried running on that.  It revealed some bugs that still need to be worked on.
I also noticed that the meanAnnualROI and meanAnnualProfit are basically
useless because they don't have a cost basis that is related at all to
the year.  All the really tell you is the average profit or average
percentage per realization, per year for all realizations.  That's not
what those terms mean so the results are misleading.  Problem is same
(though fungible of course) money used for one transaction after another
-- it's treated as separate funds which of course is inaccurate.