Did some pretty major refactoring.  Changed the way acquisitions are added.  Realization stats are now computed as acquisitions are added, even for divestments.  This means proportion of shares attributed to the acquisition are used in adding divestment stats to totals.  This affected totals for commissions and fees, as those weren't proportional before, and the entire divestment commission and fees were added to totals even if not all shares in the divestment were realized/accounted.
I also laid the foundation for calculating Set and StockAccount stats
for arbitrary date ranges.  The user will be able to set any arbitrary
date range limit, and profit and other stats will only include the
portion attributable to that date range.  In instances where a
realization is only partially within the date limit, a portion of the
value of the realization will be added to the totals, in proportion to
the ratio of the overlapping date range compared to the total
realization date range.  Actually, to be more specific, that is
calculated on a per-acquisition basis, rather than an entire
realization.

I also updated some tests and added a little documentation.