The last three commits were merged from the topic branch where I was trying to resolve the problems I'd found before.  I went with a hatchet, so it might still be useful to go through those and pull out anything worth saving.
This commit fixes a bug in $sa->numberOfTrades() where it gives the
internal number instead of the number the user is expecting.  In other
words, internally acquisitions that took place in a single trade may be
split among multiple divestments.  So some were getting counted twice or
more.  While the number it was giving may be useful for something, I
haven't found a need for it, and it should be a separate (and
differently named) method that retrieves it added back in somewhere.

I also slightly optimized how Set.pm calculates realized and unrealized
transaction counts.  And I fixed a bug in Realization.pm where the line
format values for divestments that consider proportionality were using
the wrong array index.  A little surprised I hadn't caught that before.
Added some tests around counts and updated synopsis documentation.