Revise maxCashInvested and minCashRequired
Revisit maxCashInvested.  Change calculation to match with current
understanding of what it means.  Commit now in preparation for changes
to how periodic stats for maxCashInvested are calculated.

Make related changes to minCashRequired calculation.

Change outlays to cost basis in many places as that term more accurately
reflects the meaning.

Disable warn in Transaction if it cannot calculate cash effect (as this
happens all the time in expected use cases) unless
$transaction->{verbose}.