Refactor periodic maxCashInvested
Max Cash Invested for a period should be the maximum invested during the
period.  If you sell half your account prior to period start, that the
investment in what you sold shouldn’t increase the value within the
period.

To fix, consider $total as well as $max.  The change is actually not
that big a deal… replace max in data structure with total.  Pass in
$start as well as $end to your find method.  Find the last transaction
before $start that's greater than 0, and that’s the starting total.
Compare totals after that to find max during the period.